refactor(extension): extract the workflow row and drop derived type aliases
The settings file was at the import and line caps, which forced two
obscure type aliases (Awaited<ReturnType<...>>) in place of a plain
type import. The row is self-contained, so it moves to its own file
and owns its delete confirmation and param form. WorkflowRunPrompt
now takes a name and params instead of a whole workflow.
